+++ src/gemseo_pymoo/algos/opt/_settings/unsga3_settings.py
@@ -1,0 +1,25 @@
+# Copyright 2021 IRT Saint Exupéry, https://www.irt-saintexupery.com
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU Lesser General Public
+# License version 3 as published by the Free Software Foundation.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+# Lesser General Public License for more details.
+#
+# You should have received a copy of the GNU Lesser General Public License
+# along with this program; if not, write to the Free Software Foundation,
+# Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+"""Settings for the PYMOO UNSGA3 algorithm."""
+
+from gemseo_pymoo.algos.opt._base_scaled_pymoo_algorithms_settings import (
+    BaseScaledPymooAlgorithmsSettings,
+)
+
+
+class UNSGA3Settings(BaseScaledPymooAlgorithmsSettings):
+    """The settings for the PYMOO UNSGA3 algorithm."""
+
+    _TARGET_CLASS_NAME = "PYMOO_UNSGA3"